game/serializers.py
from .models import Question, Option
from rest_framework import serializers

#Quiz Attend
class OptionSerializer(serializers.ModelSerializer):
    class Meta:
        model = Option
        fields = ['id', 'description','value']

class QuestionSerializer(serializers.ModelSerializer):
    options = OptionSerializer(many=True, read_only=True)

    class Meta:
        model = Question
        fields =['id','description','options']